+++ level0/import_l0/Dockerfile
@@ -9,11 +9,11 @@ RUN yum install postgresql-devel -y
 # Install the function's dependencies using file requirements.txt
 # from your project folder.
 
-COPY ./level0/requirements.txt  .
+COPY ./requirements.txt  .
 RUN  pip3 install -r requirements.txt --target "${LAMBDA_TASK_ROOT}"
 
 # Copy function code
-COPY ./level0/import_l0_handler/*.py ${LAMBDA_TASK_ROOT}
+COPY . ${LAMBDA_TASK_ROOT}
 
 # Set the CMD to your handler (could also be done as a parameter override outside of the Dockerfile)
-CMD [ "handler.handler" ] 
+CMD [ "handler.import_l0_handler.import_l0_handler" ]
