+++ stacks/level0_stack.py
@@ -50,7 +50,7 @@ class Level0Stack(Stack):
             self,
             "OdinSMRImportLevel0Lambda",
             code=DockerImageCode.from_image_asset(
-                ".",
+                "./level0/import_l0",
             ),
             vpc=vpc,
             vpc_subnets=vpc_subnets,
@@ -71,15 +71,15 @@ class Level0Stack(Stack):
             PolicyStatement(
                 effect=Effect.ALLOW,
                 actions=["ssm:GetParameter"],
-                resources=[f"arn:aws:ssm:*:*:parameter/{ssm_root}/*"]
+                resources=[f"arn:aws:ssm:*:*:parameter{ssm_root}/*"]
             )
         )
 
         activate_level0_lambda = Function(
             self,
             "OdinSMRLevel0Lambda",
-            code=InlineCode.from_asset("./level0/activate_l0_handler"),
-            handler="handler.handler",
+            code=InlineCode.from_asset("./level0/activate_l0"),
+            handler="handler.activate_l0_handler.activate_l0_handler",
             timeout=lambda_timeout,
             architecture=Architecture.X86_64,
             runtime=Runtime.PYTHON_3_10,
@@ -98,8 +98,8 @@ class Level0Stack(Stack):
         notify_level1_lambda = Function(
             self,
             "OdinSMRLevel1Notifier",
-            code=InlineCode.from_asset("./level0/notify_l1_handler"),
-            handler="handler.handler",
+            code=InlineCode.from_asset("./level0/notify_l1"),
+            handler="handler.notify_l1_handler.notify_l1_handler",
             timeout=lambda_timeout,
             architecture=Architecture.X86_64,
             runtime=Runtime.PYTHON_3_10,
